CS-11372: Split config.rs into submodules for Code Health 10.0
- Extract config/options.rs: ConfigOption, OPTIONS, CONFIGURABLE_TOOL_NAMES, find_option
- Extract config/validation.rs: validate_https_url, require_https, mask_if_sensitive, sensitive_env_vars
- Keep config/mod.rs: persistence, env var management, derived config
- Flatten apply_to_env bumpy road
- Resolves Low Cohesion, Bumpy Road, and String Heavy issues
- All three files now score Code Health 10.0 (was 8.41)
